feat: 자동차관련 도메인 구현
