+++ src/test/java/com/nextstep/nextstepracingcar/domain/CarTest.java
@@ -61,24 +61,5 @@ public class CarTest {
     }
 }
 
-class Car {
-
-    Name name;
-    Move move;
-
-    public Car(Name name, Move move) {
-        this.name = name;
-        this.move = move;
-    }
-
-    public Name getName() {
-        return name;
-    }
-
-    public Move getMove() {
-        return move;
-    }
-}
-
